Removed 2 projects since I am not ready to work on them.  started open gl tutorial.  updated build files to build based of name

+++ nl_lib/nl_memory.h
@@ -1,7 +1,9 @@
 #ifndef __NL_MEMORY_H__
 #define __NL_MEMORY_H__
 
-void *memory_allocate(unsigned long long size);
+#include <stddef.h>
+
+void *memory_allocate(size_t size);
 void memory_free(void* memory);
 
 #endif //__NL_MEMORY_H__
